use a bigger type for req_mem_size
The original uint16_t is enough for about 8 rings of 1024 entries.
While this is enough for current purposes it raises a question with
regard to silent 32 to 16 bit casting in operations.

To avoid this problem and to be future proof, use uint32_t for this
field in both firmware and host driver.

+++ c2x0_boot_loader/fsl_c2x0_fw.h
@@ -449,7 +449,7 @@ struct crypto_c_hs_mem {
 			u8 num_of_rps;
 			u8 max_pri;
 			u8 num_of_fwresp_rings;
-			u16 req_mem_size;
+			u32 req_mem_size;
 			u32 drv_resp_ring;
 			u32 fw_resp_ring;
 			u32 s_cntrs;
